requirements.txt
# CLI deps
tqdm==4.67.1

# base deps
pydantic==2.11.7
python-dotenv==1.1.1

# notebook + visualization
jupyter==1.1.1
ipykernel==6.30.1
plotly==6.2.0

# db
elasticsearch[async]==8.18.0

# pyesorm==0.6.12 # too opinionated

# data stuff
numpy==2.3.2
pandas==2.3.1
duckdb==1.3.2

# dev deps
pip==25.2
ruff==0.12.8
# pur: pip-update-requirements
pur==7.3.3
mypy==1.17.1
pipdeptree==2.28.0
pytest==8.4.1
pytest-xdist==3.8.0
pytest-watcher==0.4.3
pytest-asyncio==1.1.0
pytest-mypy-plugins==3.2.0
pytest-dotenv==0.5.2
